pyproject.toml
[tool.poetry]
name = "book-better"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
python-dotenv = "1.0.1"

[tool.poetry.scripts]
main = "book_better.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
